backend/requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
requests==2.31.0
pandas==2.1.1
scikit-learn==1.3.0
joblib==1.3.2
pydantic==2.7.0
python-multipart==0.0.6
google-genai==0.0.10  # if using Gemini Vision/Text via official SDK
